+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Documents/Document.cs
@@ -4,6 +4,6 @@ using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Json;
 namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Documents;
 
 [JsonConverter(typeof(DocumentConverter<Document, object>))]
-public record Document: Document<object>
+public record Document: DocumentBase<object>
 {
 }
