+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/GeoPoint.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json;
 using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Json;
 using SearchGeoPoint = VirtoCommerce.SearchModule.Core.Model.GeoPoint;
 
-namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api;
+namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query;
 
 [JsonConverter(typeof(GeoPointConverter))]
 public class GeoPoint: SearchGeoPoint
