+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/MeasurementUnit.cs
@@ -1,4 +1,4 @@
-namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api;
+namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query;
 
 public enum MeasurementUnit
 {
