+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Result/SearchResultDocument.cs
@@ -3,8 +3,8 @@ using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Json;
 
 namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Result;
 
-[JsonConverter(typeof(DocumentConverter<Document, FieldValue>))]
-public record Document: Document<FieldValue>
+[JsonConverter(typeof(DocumentConverter<SearchResultDocument, FieldValue>))]
+public record SearchResultDocument: DocumentBase<FieldValue>
 {
     [JsonProperty("_meta")]
     public virtual DocumentMetadata Meta { get; init; }
