+++ src/VirtoCommerce.ElasticAppSearch.Core/Services/Converters/IDocumentConverter.cs
@@ -1,5 +1,6 @@
 using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Documents;
 using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Schema;
+using VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Result;
 using VirtoCommerce.SearchModule.Core.Model;
 
 namespace VirtoCommerce.ElasticAppSearch.Core.Services.Converters;
@@ -8,5 +9,5 @@ public interface IDocumentConverter
 {
     (Document, Schema) ToProviderDocument(IndexDocument indexDocument);
 
-    SearchDocument ToSearchDocument(Models.Api.Search.Result.Document searchResultDocument);
+    SearchDocument ToSearchDocument(SearchResultDocument searchResultDocument);
 }
